refac - queryString converter moved
