+++ src/main/java/org/sparta/foodordermanagementservice/common/WebConfig.java
@@ -1,7 +1,7 @@
 package org.sparta.foodordermanagementservice.common;
 
 import org.sparta.foodordermanagementservice.dto.request.OrderListRequestCondition;
-import org.sparta.foodordermanagementservice.dto.request.SortedByConverter;
+import org.sparta.foodordermanagementservice.common.utils.SortedByConverter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
