+++ src/main/java/org/sparta/foodordermanagementservice/common/utils/SortedByConverter.java
@@ -1,6 +1,7 @@
-package org.sparta.foodordermanagementservice.dto.request;
+package org.sparta.foodordermanagementservice.common.utils;
 
 import lombok.NonNull;
+import org.sparta.foodordermanagementservice.dto.request.SortedBy;
 import org.springframework.core.convert.converter.Converter;
 
 public class SortedByConverter implements Converter<String, SortedBy> {
